Refactoring: Reusable module for downloading FDB binaries
